Populated the details page table with dummy data.
